Customized error message in case of invalid submission of form and started with styling

+++ src/components/ClientListItem.js
@@ -11,7 +11,7 @@ export const ClientListItem = ({
   phone,
   gender = 'N/A',
   address,
-  status = 'N/A',
+  priority = 'N/A',
   field = 'N/A',
   memberType = 'N/A',
   createdAt,
@@ -19,10 +19,33 @@ export const ClientListItem = ({
   id,
   dispatch
 }) => {
+  let priorityRating;
+  switch (parseInt(priority)) {
+    case 5:
+      priorityRating = 'Very High';
+      break;
+    case 4:
+      priorityRating = 'High';
+      break;
+    case 3:
+      priorityRating = 'Medium';
+      break;
+    case 2:
+      priorityRating = 'Low';
+      break;
+    case 1:
+      priorityRating = 'Dormant';
+      break;
+    default:
+      priorityRating = 'N/A';
+  }
+  console.log(gender);
+
   return (
     <div className="clientListItem">
       <Link to={`/view/${id}`} className="clientListItem__name">
         <p>
+          {gender === 'N/A' ? '' : gender === 'M' ? 'Mr. ' : 'Ms. '}
           {firstName} {lastName}
         </p>
       </Link>
@@ -31,16 +54,18 @@ export const ClientListItem = ({
       <p> {phone}</p>
       <p> {gender}</p>
       <p> {address}</p>
-      <p> {status}</p>
+      <p> {priorityRating}</p>
       <p> {field || 'No Record'}</p>
       <p> {memberType}</p>
       <p> {moment(createdAt).format('DD/MM/YYYY')}</p>
       <p> {moment(lastCommunication).format('DD/MM/YYYY')}</p>
 
       <a
-        className="clientListItem__button--one"
+        className="clientListItem__butt on--one"
         onClick={() => {
-          dispatch(startRemoveClient({ id }));
+          if (confirm(`Are you sure you wanna delete ${firstName}'s record?`)) {
+            dispatch(startRemoveClient({ id }));
+          }
         }}
       >
         Remove
